Major changes: Added fromStrems and suing ConcurrentHashMap in Memtable

+++ event-store/src/main/java/io/joshworks/fstore/es/index/Index.java
@@ -17,6 +17,6 @@ public interface Index extends Closeable, Iterable<IndexEntry> {
 
     int version(long stream);
 
-    long size();
+    int size();
 
 }
